1. tests/unit/test_plot_paper_module.py 추가   2. tests/unit/test_plot_diagnostic_module.py 추가   1. ruff check . 통과   2. 신규 타깃 테스트 통과: 4 passed   3. 전체 게이트 통과: 145 passed, 1 skipped   4. 전체 커버리지: 66.04%   5. 모듈 커버리지 개선:       - collector/plot_paper.py: 87%       - collector/plot_diagnostic.py: 91%       - python -m edge.edge_daemon --help       - python -m collector.collector --help       - python -m collector.analyze --help       - python -m experiments.run_scenarios --help

+++ edge/sensors/mic_rms.py
@@ -209,7 +209,8 @@ class MicRMS:
             rms = float(np.sqrt(np.mean(x * x)))
             dbfs = 20.0 * math.log10(max(rms, eps)) + self.db_offset
             # clipping 비율
-            clip_ratio = float(np.mean(np.abs(frame) >= thr_val))
+            # Promote to int32 first to avoid int16 overflow on abs(-32768).
+            clip_ratio = float(np.mean(np.abs(frame.astype(np.int32)) >= thr_val))
             s = Sample(ts_ns=ts_ns, seq=self.seq, dbfs=dbfs, clip_ratio=clip_ratio)
             self.seq += 1
             yield s
